[feat] 조회 재행동 시 횟수에 따라 가중치 차등 업데이트 처리

+++ src/main/java/com/dto/project/domain/weighting/config/WeightingProperties.java
@@ -18,7 +18,7 @@ import java.util.Map;
 public class WeightingProperties {
 
     //로그 점수 및 계수 저장
-    private Map<WeightLogType, Integer> actionWeight = new EnumMap<>(WeightLogType.class);
+    private Map<WeightLogType, Double> actionWeight = new EnumMap<>(WeightLogType.class);
     private Map<MetaTagType, Double>   metaTagCoefficient  = new EnumMap<>(MetaTagType.class);
 
     private double occupationBonus;
